few merge fix

+++ upmem/src/dpu/dpu_embedding.c
@@ -20,8 +20,8 @@ __host uint32_t counter_all, counter_init;
 
 BARRIER_INIT(my_barrier, NR_TASKLETS);
 
-uint32_t indices_len;
-uint32_t nr_batches;
+uint64_t indices_len;
+uint64_t nr_batches;
 __dma_aligned struct query_len lengths;
 __dma_aligned uint32_t indices[MAX_INDEX_PER_BATCH * MAX_NR_BATCHES];
 __dma_aligned uint32_t offsets[MAX_NR_BATCHES];
@@ -38,7 +38,7 @@ main() {
         indices_len = lengths.indices_len;
         nr_batches = lengths.nr_batches;
 
-        uint32_t copied_indices = 0;
+        uint64_t copied_indices = 0;
         while (copied_indices < indices_len) {
             mram_read(&input_indices[copied_indices], &indices[copied_indices],
                       ALIGN(MIN(2048, (indices_len - copied_indices) * sizeof(uint32_t)), 8));
@@ -53,17 +53,17 @@ main() {
     }
     barrier_wait(&my_barrier);
 
-    uint32_t batch_lot_size = 2048 / sizeof(*results);
-    uint32_t nr_batch_lots = (nr_batches / batch_lot_size) + 1;
-    uint32_t batch_end = 0;
-    for (uint32_t batch_lot=0; batch_lot < nr_batch_lots; batch_lot++) {
-        uint32_t batch_start = batch_end;
+    uint64_t batch_lot_size = 2048 / sizeof(*results);
+    uint64_t nr_batch_lots = (nr_batches / batch_lot_size) + 1;
+    uint64_t batch_end = 0;
+    for (uint64_t batch_lot=0; batch_lot < nr_batch_lots; batch_lot++) {
+        uint64_t batch_start = batch_end;
         batch_end = MIN(batch_start + batch_lot_size, nr_batches);
-        for (uint32_t i = me() + batch_start; i < batch_end; i += NR_TASKLETS) {
+        for (uint64_t i = me() + batch_start; i < batch_end; i += NR_TASKLETS) {
             tmp_results[i] = 0;
-            uint32_t upper_bound = i == nr_batches - 1 ? indices_len : offsets[i + 1];
-            for (uint32_t indices_ptr = offsets[i]; indices_ptr < upper_bound; indices_ptr++) {
-                uint32_t ind = indices[indices_ptr];
+            uint64_t upper_bound = i == nr_batches - 1 ? indices_len : offsets[i + 1];
+            for (uint64_t indices_ptr = offsets[i]; indices_ptr < upper_bound; indices_ptr++) {
+                uint64_t ind = indices[indices_ptr];
                 tmp_results[i] += emb_data[ind];
             }
         }
